refactor: swap test scripts
