+++ scripts/tests/swap-permit.ts
@@ -1,29 +1,15 @@
-import axios from "axios";
-import dotenv from "dotenv";
-import {
-  filterTestCases,
-  MIN_OUTPUT_CASES,
-  EXACT_OUTPUT_CASES,
-  SWAP_API_BASE_URL,
-} from "./_swap-utils";
-dotenv.config();
+import { swap } from "./_swap-utils";
 
-/**
- * Manual test script for the swap API. Should be converted to a proper test suite.
- */
-async function swap() {
-  const filterString = process.argv[2];
-  const testCases = [...MIN_OUTPUT_CASES, ...EXACT_OUTPUT_CASES];
-  const filteredTestCases = filterTestCases(testCases, filterString);
-  for (const testCase of filteredTestCases) {
-    console.log("\nTest case:", testCase.labels.join(" "));
-    const response = await axios.get(`${SWAP_API_BASE_URL}/api/swap/permit`, {
-      params: testCase.params,
-    });
-    console.log(response.data);
-  }
+async function swapWithPermit() {
+  console.log("Swapping with permit...");
+  await swap();
 }
 
-swap()
+swapWithPermit()
   .then(() => console.log("Done"))
-  .catch(console.error);
+  .catch((e) => {
+    console.error(e);
+    if (e.response?.data) {
+      console.log("Tx for debug sim:", e.response.data.transaction);
+    }
+  });
